Test cases and logging

+++ src/main/java/com/example/springdatademo/service/UserServiceImpl.java
@@ -4,6 +4,9 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
+import org.apache.logging.log4j.util.Strings;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -13,16 +16,18 @@ import org.springframework.stereotype.Component;
 
 import com.example.springdatademo.dto.UserDto;
 import com.example.springdatademo.entity.User;
+import com.example.springdatademo.exception.UserNotFoundException;
 import com.example.springdatademo.repository.UserRepository;
 
 @Component
 public class UserServiceImpl implements UserService{
+	Logger logger = LoggerFactory.getLogger(UserServiceImpl.class);
+	
 	@Autowired
 	UserRepository repository;
 
 	@Override
 	public List<User> getUsers() {
-		
 		
 		return repository.findAll();
 	}
@@ -33,9 +38,15 @@ public class UserServiceImpl implements UserService{
 	}
 
 	@Override
-	public void deleteUser(Integer id) {
+	public void deleteUser(Integer id) throws UserNotFoundException {
+		logger.info("#####START:UserServiceImpl.deleteUser()#####");
 		Optional<User> demoUser = repository.findById(id);
-		repository.delete(demoUser.get());		
+		if(!demoUser.isPresent()){
+			logger.error("No User with this Id exists ");
+			throw new UserNotFoundException("No User with this Id exists ");
+		}
+		repository.delete(demoUser.get());
+		logger.info("#####END:UserServiceImpl.deleteUser()#####");
 	}
 
 	@Override
@@ -45,8 +56,15 @@ public class UserServiceImpl implements UserService{
 	}
 
 	@Override
-	public User getUserById(Integer id) {
-		return repository.findById(id).get();
+	public User getUserById(Integer id) throws UserNotFoundException {
+		logger.info("#####START:UserServiceImpl.getUserById()#####");
+		Optional<User> optional = repository.findById(id);
+		if(!optional.isPresent()){
+			logger.error("No User with this Id exists ");
+			throw new UserNotFoundException("No User with this Id exists ");
+		}
+		logger.info("#####END:UserServiceImpl.getUserById()#####");
+		return optional.get();
 	}
 	
 	@Override
@@ -55,11 +73,24 @@ public class UserServiceImpl implements UserService{
 	}
 
 	@Override
-	public List<User> getUsers(Integer pageNo, Integer pageSize, String sortBy) {
+	public List<User> getUsers(Integer pageNo, Integer pageSize, Integer yearsOfExperienceMin,
+			Integer yearsOfExperienceMax, String sortBy) {
+		Pageable page;
+		Page<User> pagedResult;
+		if (Strings.isBlank(sortBy))
+			page = PageRequest.of(pageNo, pageSize);
+		else
+			page = PageRequest.of(pageNo, pageSize, Sort.by(sortBy));
 
-		Pageable pageable = PageRequest.of(pageNo, pageSize,Sort.by(sortBy));
 
-		Page<User> pagedResult = repository.findAll(pageable);
+		if (yearsOfExperienceMin != null && yearsOfExperienceMax != null) {
+
+			pagedResult = repository.findByYearsOfExperienceGreaterThanEqualAndYearsOfExperienceLessThanEqual(
+					yearsOfExperienceMin, yearsOfExperienceMax, page);
+
+		} else {
+			pagedResult = repository.findAll(page);
+		}
 
 		if (pagedResult.hasContent()) {
 			return pagedResult.getContent();
@@ -67,9 +98,6 @@ public class UserServiceImpl implements UserService{
 			return new ArrayList<User>();
 		}
 
-	
 	}
-
-
 
 }
